+++ components/float-button/index.ts
@@ -1,7 +1,17 @@
-export { FloatButton } from './FloatButton'
-export { FloatButtonGroup } from './FloatButtonGroup'
-export { FloatButtonBackTop } from './FloatButtonBackTop'
-export { default } from './FloatButton'
+import { FloatButton as InternalFloatButton } from './FloatButton'
+import { FloatButtonGroup } from './FloatButtonGroup'
+import { FloatButtonBackTop } from './FloatButtonBackTop'
+
+type FloatButtonCompound = typeof InternalFloatButton & {
+  Group: typeof FloatButtonGroup
+  BackTop: typeof FloatButtonBackTop
+}
+
+const FloatButton = InternalFloatButton as FloatButtonCompound
+FloatButton.Group = FloatButtonGroup
+FloatButton.BackTop = FloatButtonBackTop
+
+export { FloatButton, FloatButtonGroup, FloatButtonBackTop }
 export type {
   FloatButtonProps,
   FloatButtonGroupProps,